requirements.txt
pyyaml
requests
beautifulsoup4
# Optional for future enhancements
# tweepy  # Twitter API
# yfinance  # Market data
# pandas
# numpy
